refactor main class create a interface for exchanges
