+++ exchanges/constants.go
@@ -1,0 +1,5 @@
+package exchanges
+
+const (
+	pancakeswap = "https://api.pancakeswap.info/api/v2/tokens/"
+)
