+++ go.mod
@@ -1,0 +1,3 @@
+module github.com/kimagliardi/coinTracker
+
+go 1.16
